+++ src/main/java/kitchenpos/products/tobe/domain/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.tobe.domain;
 
 import java.util.List;
 import java.util.Optional;
